refactor: TMS-2208 add tenant code cleanup (#20)
* refactor: TMS-2208 clean up create tenant code and ui

Fix the notifications so they can be dismissed by clicking anywhere, not just the close X.

* refactor: TMS-2208 clean up create tenant code and ui

Fix the navigation bar to provide horizontal space between the buttons.

* refactor: TMS-2208 clean up create tenant code and ui

Fix the console error handling to provide more meaningful error messages

* refactor: TMS-2208 clean up create tenant code and ui

Rename the CreateTenantDialog

* refactor: TMS-2208 clean up create tenant code and ui

Major refactor to make the CreateTenantDialog "dumb" - it now uses props and events, and does not know anything about the state. All state stuff is moved to the "smart" view.

* refactor: TMS-2208 clean up create tenant code and ui

Removed the generation of notifications from the store. The store should be concerned with state, not UI.

* refactor: TMS-2208 clean up create tenant code and ui

Big refactor of the dialog to change it from Vue2 style to Vue3.