fixed frontend build issues
frontend had a lot of build issues

there was a lot of confusion as to what was doing what because of the usage of the word tenant/tenancy/tenants/tenancies, all of which mean different things. i updated everything to be a bit more clear, however it's still using unclear information from the backend.