fix: TMS-2208 notification (#35)
* fix: TMS-2208 get tenants notification

If getting tenants for the user failed, it was silently just displaying no tenants. Fixed the onMounted to wait for the store and produce a notification if needed.

* docs: TMS-2208 tenant service jsdocs

Added a few jsdocs for the tenant service methods that are complete and being used. Others to come.

* refactor: TMS-2208 move auth into store

Changed the keycloak service to be an authentication store.

* fix: TMS-2208 add axios auth interceptor

Updated the axios use to include the bearer token on requests.