fix: TMS-2208 tenant validation
The dialog now will clear the fields before it is displayed. This is to prevent two things: firstly that if the user enters data, cancels, and displays the dialog again, the data is not shown. Second that if the data is cleared before it pops down then validation happens and momentarily looks ugly.