fix: TMS-2208 ux updates (#48)
* fix: TMS-2208 ux tweaks to match wireframes

Was working from stale wireframes:
- validate the dialog when it's displayed so that the user knows what to fill in
- display the duplicate name message in the dialog, not a notification
- clear duplicate name message when either field is changed
- change BC Ministries to Ministries/Organizations

* fix: TMS-2208 ux tweaks to match wireframes

Added primary and secondary button components. Still need L&F changes.

* fix: TMS-2208 ux tweaks to match wireframes

First commit of the Vuetify theme. Colour will need to be fixed once the design system is available.

* fix: TMS-2208 ux tweaks to match wireframes

Added some space in the dialog

* fix: TMS-2208 ux tweaks to match wireframes

Removed the icon for the multitenancy info link.

* fix: TMS-2208 ux tweaks to match wireframes

Bump the disabled opacity for contrast, but may not be enough - slippery.

* fix: TMS-2208 ux tweaks to match wireframes

Replaced the hokey requestAnimationFrame with nextTick. Renamed the update-name event and also have it clearing the validation message when ministry changes.

* fix: TMS-2208 ux tweaks to match wireframes

Rework the dialog to strictly emit the entered values - the parent cannot pass them in.

* fix: TMS-2208 ux tweaks to match wireframes

Final tweaks and copilot suggestions for ARIA and TypeScript.