refactor: TMS-2210 split manage tenant component (#51)
* refactor: TMS-2210 manage tenant conversion

Converted one big component into a view and multiple presentational components. This is only the first piece of a lot of work.

* fix: TMS-2210 replace tenancy with tenant

* feat: TMS-2210 add created by and date

In the header display the date the tenant was created, and who created it.

* feat: TMS-2210 add fields for details

In the tenant details also display the tenant name and description (new).

* refactor: TMS-2210 basic business logic

Got the edit button working and calling the service, which returns a dummy response until the route is completed.

* refactor: TMS-2210 emit function name

Changed the emit function name to make it consistent with other components..

* refactor: TMS-2210 delete unused files

Now that refactoring is almost complete, the original files can be removed.