fix: TMS-2210 nav breadcrumbs and other ux changes (#52)
* fix: TMS-2210 persistent navigation highlights

The navigation bar was set up to only highlight when clicked. Now the tenant button will stay highlighted when viewing the details for tenant.

* refactor: TMS-2210 new breadcrumb bar

Created a breadcrumb bar component. Mostly to fix the divider that is used but also to ensure consistency in future views.

* refactor: TMS-2210 rename custom buttons

Renamed the custom button components so that they're less like vuetify names.

* fix: TMS-2210 move edit menu to detail

The edit menu was in the tenant header, but it needs to be in the detail itself. Duplicated it for now, and will convert the old one later.

* fix: TMS-2210 show/hide tenant detail

Changed the header to remove the edit/delete menu and instead have a button that hides or shows the tenant detail.

* fix: TMS-2210 force load of tenant

Changed the management view to load the tenant as it might be stale in the store if another user changed it.