fix: TMS-2210 users missing when viewing tenant (#54)
* fix: TMS-2210 handle missing users

Add some protection in case there are no users for a tenant. This should not happen but it's good to fail gracefully.

* fix: TMS-2210 handle missing users

Add some protection in case there are no users for a tenant. This should not happen but it's good to fail gracefully.

* fix: TMS-2210 extend notification time

Notifications were only displayed for three seconds, so extended it to ten seconds.

* fix: TMS-2210 handle missing users

Add some protection in case there are no users for a tenant. This should not happen but it's good to fail gracefully.

* fix: TMS-2210 handle missing users

Added the API call flag that causes the users to be retrieved.