fix: TMS-2210 update ux to match wireframe (#57)
* fix: TMS-2210 ux wireframe updates

Doing some updates to match the wireframe. Changed the tenant create date to be a string, so that timezone issues can be avoided.

* refactor: tenant createdDate rename

Renamed the tenant createDateTime field from the API to createdDate in the frontend. Hopefully will reduce confusion.

* refactor: emit change for sonar

Refactored the emit to make sonarqube happy.

* refactor: tenant header tweaks

Fixed up the header to have everything vertically aligned, no wrapping of date values, and a blue-ish background colour.

* refactor: tenant tabs tweaks

Removed the extra tab and changed it to not open a tab by default - the user has to click a tab to display it.

* refactor: tenant header tweaks

Reworked the header to keep the non-title fields snug on the right side.

* refactor: tenant list spacing

Rework the spacing to have it controlled by the app so that each view just has to make sure it has no outer spacing

* refactor: application spacing

Shift the application spacing up so that only the parents should need to worry about spacing.

* fix: tenant details spacing

Fix up all the spacing in the view of the tenant details to be closer to the wireframe.