fix: TMS-2210 tenant form updates (#58)
* fix: TMS-2210 tenant details disabled

In the tenant details changed the fields from readonly to disabled, since they are very different behaviours (focus, hover, etc).

* fix: TMS-2210 form component updates

Changed the form components from readonly to disabled. Switched description to text area. Added cancel button. Consolidated validation so that save button can be disabled if not valid.

* feat: TMS-2210 form cancel icon

Added an X icon that replaces the kebab icon when the form is being edited.