fix: TMS-2210 layout and spacing changes (#59)
* fix: TMS-2210 additional ui updates

Added some extra space above the tabs. Alphabetized up the imports.

* fix: TMS-2210 alphabetized imports

Did some cleanup of the imports for consistency and readability.

* build: TMS-2210 update vuetify

Updated Vuetify to latest in the hope of fixing a bug but no such luck. Accidentally included some other work but it's needed anyway!

* fix: TMS-2210 rework layout for tenant

Redid the layout so that the header and details are responsive when the window resizes.