ci: updating ci/cd configuration (#24)
* Update .deployer.yml

updated oc-runner from 1.0.0 to 1.2.2

* Create Dockerfile

created a frontend Dockerfile to be built

* Update Dockerfile

fixed github advanced security issues:

image user is not root
healthcheck defined

* Update package.json

added missing coverage tests

* switch nginx to caddyfile

switched from nginx to caddyfile as i was getting permission errors

* Update Dockerfile

added install of packages so we have the run-p command

* formatting

caddyfile uses spaces as indentation
allowed dockerfile to overrwrite the formatting

* update backend configuration

configured the readiness and liveness probe path, removed port 3000
set the correct targetPort for our purposes.

fixed the docker-compose in the root for the backend

* Updated deployment

disabled pgBackRest and pgmonitor as it's not important for the MVP
added the missing http port
removed prometheus metrics from frontend since we don't use that
added health route for port 3001 for frontend caddyfile and removed metrics

* added environment variables

added environment variables for the backend

* Update pr-open.yml

was passing these to deployer.yml which didn't use them

* Update pr-open.yml

test to see if setting strings works

* Update pr-open.yml

created a job to update the secrets, so it should be able to set the proper client id and client secret securely

* Update pr-open.yml

updated the correct client_id secret

* added backend ingress

added an ingress for the backend

* Update service.yaml

changing the targetPort to the values configured which should be 4144

* Update knp.yaml

set the targetPort so front end can access the back end

added an ingress network policy for the backend

* Update pr-open.yml

removed comment

* added frontend configuration

added environment variables and secrets for the front end to run.

* update backend environment variables

The backend environment variables BC_GOV_SSO_CLIENT_ID was changed to BCGOV_SSO_API_CLIENT_ID and BCGOV_SSO_CLIENT_SECRET was changed to BCGOV_SSO_API_CLIENT_SECRET

* Update .env.sample

removed other instances of VITE_ALLOWED_HOSTS as it's not a used env var

* fix: configmap usage

instead of using secrets, switched to configmap

the frontend container should mount a config file called default.json containing the values we are looking for.

* Update deployment.yaml

removed secret env reference

* Update pr-open.yml

Creates the config before building because patching didn't seem to work

* Update pr-open.yml

fixed the configmap that should be updated

* Update pr-open.yml

moved the configmap changes back into postbuild because i think the helm chart configmap that is generated just overwrites what is patched in the github action.

* Update pr-open.yml

removed trailing comma
attempting to delete the configmap and create it, then delete the pods to try to force the configmap to be refreshed

* updated reverse proxy

reverse proxy was going to a missing environment variable, and needed to separate the reverse proxy from the frontend and caddy

* Update Caddyfile

removed the uri subdirective as it does not seem to work

* Update Caddyfile

switched http transport to tls

* removed tls from caddy

caddyfile had tls which isn't needed.. needed to set the deployment to use https

* Update Caddyfile

match requests with the pattern /api/v1/*
remove only the /api prefix, leaving the /v1/* part intact
forward the request to the backend with the correct path

* Update Caddyfile

trying another way to handle the reverse proxy.
previous method was outdated

* Update Caddyfile

typo, used ${} instead of {$}

* added actual client id

temporarily adding client id until solution can be found

* feat: update-secrets workflow

created a workflow to update the secrets or variables after the build has been completed.

* Update Caddyfile

Explicitly handle /config/* paths
Servers file from the root instead of /srv for these paths
Excludes /config* from SPA routing

* Update config.service.ts

replaced console logging functions with logger

* added file_server to caddyfile

added file_server to caddyfile and added permissions to update_secrets workflow

* Update Caddyfile

added /config to the exclusion list for the spa_router

* Update .update-secrets.yml

replaced baseUrl with the /api/v1 route

* removed envConfig

i suspect envConfig is what is causing the values to be overridden

* Update config.service.ts

debug logging to see what the loaded config is

* fix: config loading

the config should be loaded before keycloak is initialized

* Update authenticated.axios.ts

fixed issue where keycloak is possibly null

* Update Dockerfile

added health check for frontend

* Update values.yaml

trying to disable backups for crunchy

* fixed cicd

prod deployment would have deployed with the test name instead of prod

* fix: cicd tests

cicd tests were trying to run test:cov
backend npm script was created for test:cov
frontend npm package was missing dependency