feat: TMS-2209 add user to tenant (#76)
* feat: TMS-2209 add tenant user

Huge rewrite of the user search to put it into its own component. Create a new store and refactor the service. Work in progress.

* fix: TMS-2209 user info layout

Updated the layout of the user search info to not take up the whole width of the screen and to be vertically aligned.

* feat: TMS-2209 map user search

Added a new method to the model to map user search data to a User.

* refactor: TMS-2209 tenant removal

Removed the tenant lookup function from the service, as it's now supplied by the tenant service.

* feat: TMS-2209 map user search

Now using the response mapping to create a proper User. Removed the tenant code as it's not needed anymore.

* fix: TMS-2209 user single select

Set up the data table to only allow single selections.

* fix: TMS-2209 button alignment

Hide the details (validation messages) so that the button aligns correctly.

* fix: TMS-2209 add user updates

Big update of the add user to tenant functionality. Still needs work but it will do for now.