fix: TMS-2209 ux tweaks for adding user to tenant (#92)
* refactor(ui): TMS-2209 emits consistency

Made the emits consistent and in line with the Vue recommendations.

* refactor(ui): TMS-2209 tenant add user

Refactored the add user to move the store access into the view. The presentational components should not be accessing the store.

* refactor(ui): TMS-2209 import ordering

* refactor(ui): TMS-2209 remove searchAvailableUsers

The tenant store previously had the searchAvailableUsers, but that has been moved to the user store and is not longer needed.

* refactor(ui): TMS-2209 reorder exports

* refactor(ui): TMS-2209 store simplification

Cleaned up the tenant store to remove the functionality that is no longer being used - users and roles are no longer fetched individually, they are grabbed when the tenant is grabbed.

* fix(ui): TMS-2209 update user list

Although users were being successfully added. the store was not being updated and so the user was not being shown.

* feat(ui): TMS-2209 allow multiple roles

The user should be add-able with multiple roles. Use a set of checkboxes for them.

* feat(ui): TMS-2209 allow multiple roles

Now that the form allows multiple roles, set up everything else so that those roles are passed to the API.