fix: TMS-2209 additional UI updates to match wireframes (#93)
* feat(ui): TMS-2209 add tenant users header

Add the "Tenant Users" header that is in the wireframes but not the poc.

* feat(ui): TMS-2209 sort users by name

Sort the tenant users by display name, which means that they're sorted by surname.

* revert(ui): TMS-2209 remove debug log

Remove a debugging line that was added in previous PR.

* revert(ui): TMS-2209 change console to logger

Updated the view to use the logger class and not the console.

* feat(ui): TMS-2209 add user instructions

Add the instructions to look up the user.

* feat(ui): TMS-2209 duplicate user error

If the API returns a 409 when adding the user, display an appropriate error message.

* fix(ui): TMS-2209 duplicate user error

If the API returns a 409 when adding the user, also clear the search results so that they are not displayed if the user chooses to search again.

* fix(ui): TMS-2209 wireframe stylings

Update text and stylings to more closely resemble the wireframes.

* feat(ui): TMS-2209 add user cancel button

Added the cancel button when the add user form is displayed, and hide the main button above.