feat: TMS-2287 update tenant (#95)
* feat(ui): TMS-2287 update tenant

Now that the update route exists, use it to update the tenant after editing.

* fix(ui): TMS-2287 update tenant errors

Now that the API route exists, add API errors handling. Changed the watch to a onMounted to be more idiomatic.

* fix(ui): TMS-2287 tenant management id

Removed duplicate onMounted and removed check that id exists - the routes should guarantee this.

* refactor(ui): TMS-2287 tenant management cleanup

Cleaned up the view to make the code less scattered. The first of a lot of much-needed work.

* fix(ui): TMS-2287 tenant update on save

Fix up the store so that the tenant is updated after it is saved.

* fix(ui): TMS-2287 handle duplicate tenant

Added the code to handle the duplicate tenant name error. Still more to be done.

* fix(ui): TMS-2287 max length for description

Set the maximum length for the description, made it a required field, and turned on the counter.