Reduce data volume storage to 1Gi in values.yml (#100)
Updated the dataVolumeClaimSpec storage from 3Gi to 1Gi in the Crunchy chart values file. This change may help optimize resource usage or adapt to new storage requirements.

We currently use 118.4 MiB of 3 GiB for the pgdata and 161.1 MiB out of 300 MiB for pgwal.

It's failing to allocate enough resources for the second replica because we can't request anymore.