fix: Update pr-close workflow and remove cleanup_db job (#104)
* Update pr-close workflow and remove cleanup_db job

Updated the referenced workflow version for cleanup and image promotion. Removed the cleanup_db job for removing DB users from Crunchy, this should be handled by the updated version of pr-close workflow.

* Fix values_file selection logic in deployer workflow

Corrects the conditional for selecting the values_file in the deployer GitHub Actions workflow to use the 'environment' input instead of 'tag'. This ensures the correct values file is used based on the deployment environment.