feat(ui): TMS-2386 use design tokens (#108)
* feat(ui): TMS-2386 use design tokens

Switch the Vuetify theme to use the bcgov design tokens. Will need to adjust these component by component, but the start is to start pulling them from the design components package.

* feat(ui): TMS-2386 use design tokens

Stripped down the vuetify configuration and now it's only defining things we'll use. More may be added but it's good for now.