refactor: TMS-2508 buttons and eslint rules (#122)
* refactor(ui): TMS-2508 move margin defs

Move the primary and secondary margins into their containers - where needed - as they do not always need to have margins.

* fix(ui): TMS-2508 remove console.log

* refactor(ui): TMS-2508 use primary button

Update the user search to use the primary button instead of the raw v-btn styled as primary.

* refactor(ui): TMS-2508 remove store notification

Remove the creation of a notification in the store. Only the views should be doing UI things - there were actually two notifications being displayed because of this.

* refactor(ui): TMS-2508 convert btn to fab

Converted a raw v-btn to the new Floating Action Button component.

* feat: TMS-2508 add vue linting rule

Update the linter to add rules and start to fix the results.

* refactor(ui): TMS-2508 linter fixes

Huge number of linter fixes now that it's being more strict. This is the first round, with many more improvements to come.

* refactor(ui): TMS-2508 linter a13n fixes

Uncommented the alphabetization rules and did all the fixes.

* feat(ui): TMS-2508 more linter rules

Added some more linter rules so that we're more strict about the allowed code.