refactor: TMS-2508 sonar fixes and view updates (#123)
* refactor(ui): TMS-2508 view cleanups

Before starting on the view cleanups, will fix some sonar issues introduced when the configuration work was done.

* refactor(ui): TMS-2508 config service

Did some minor alphabetization.

* feat(ui): TMS-2508 api error type guards

Sonar complained about duplicated code on the last PR. Added some type guards for 400 and 409 errors to consolidate some of this code.

* refactor(ui): TMS-2508 change roles to view

The linter was complaining that Roles isn't a valid name so renamed it to RolesView and set it up as such.

* refactor(ui): TMS-2508 header / store separation

Refactored the header so that it no longer directly accesses the store. The header now uses the proper props and events, and the App will handle the store use.

* refactor(ui): TMS-2508 tenant list renames

Moved the tenant list and card components into the tenant directory. Renamed tenant card to tenant list card to make it clearer where it's used.

* refactor(ui): TMS-2508 tenant create rename

Moved the tenant create dialog into the tenant directory and renamed.

* fix(ui): TMS-2508 shrink user icon

Shrank the user icon so that although it's a different size (by definition) from the logout icon, it appears the same size on the screen since it isn't in a button.

* refactor(ui): TMS-2508 tenant card colours

Updated the tenant card to use design token colours for the background and the links. Also simplified the component by removing code to handle the error case of a tenant having no owners (silently fails instead of a notification).

* refactor(ui): TMS-2508 navigation colour

Updated the navigation background colour to use the design tokens.

* fix(ui): TMS-2508 remove aria attrs

A couple of components had ARIA attributes - removed them until a solid ARIA plan is developed.

* refactor(ui): TMS-2508 rename simpledialog

Renamed DialogBox to SimpleDialog.

* fix(ui): TMS-2508 global styles for links

* feat(ui): TMS-2508 disable input styles

Added configuration so that colours for disabled inputs can be controlled. Using the design token defaults but probably needs more contrast.

* refactor(ui): TMS-2508 base secure view

Renamed the BaseSecure to be a view since it calls the store. Can be decomposed later if it gets too complex.