refactor: TMS-2508 design tokens for fonts and colours (#126)
* refactor(ui): TMS-2508 tenant detail colour

Set up the colours for the tenant header and details to use the design tokens.

* refactor(ui): TMS-2508 use prefixed tokens

Switch to the design tokens with the "bcds" prefix. This will serve as a reminder of the source of the tokens.

* refactor(ui): TMS-2508 design system typography

Set up the typography from the design system and set up headers and paragraphs. More to come, as needed.