refactor(ui): TMS-2509 add ssouser object and cleanup (#128)
* refactor(ui): TMS-2509 rbac setup

Started the RBAC setup by doing better type safety in the models. Added a getFirstOwner method to the tenant to make things easier.

* refactor(ui): TMS-2509 user rewrite

Completely redid the user model to undo a bad design decision. The model now tracks the API data, with a user having a ssoUser model within it.

* fix(ui): TMS-2509 remove any types

After the previous cleanup it's now possible to remove the "any" typed arguments that can now have stronger types.

* fix(ui): TMS-2509 remove component logging

Removed logging from the components - their responsibility is the user, not the logs.

* fix(ui): TMS-2509 user search sorting

The user to user + ssouser work broke the sort for the user search data table. Also through in some bug fixes to clear the search results.