refactor(ui): TMS-2509 rbac and code cleanup (#129)
* refactor(ui): TMS-2509 update tenant todo

Now that the models are becoming more strict it was possible to remove a todo and tighten up the store code.

* refactor(ui): TMS-2509 add tenant details

Added a new TenantDetails type that the editing component can use for a subset of the tenant fields.

* refactor(ui): TMS-2509 tenant loading

The view was over-complicated in that it was trying to deal with whether the tenant had loaded yet or not. Put the template in a wrapper that handles the loading, and used a computed value to prevent null checks everywhere.

* refactor(ui): TMS-2509 tenant simplifications

More simplifications: remove extra references in the view, and make the handler arguments easier to work with. This reduces both code and API calls.

* refactor(ui): TMS-2509 use store loading

Simplify the view to use the loading variable in the store, rather than manage it again.

* fix(ui): TMS-2509 loading centering

The loading text was not quite centered and looked a little funny.

* fix(ui): TMS-2509 remove search type

There was a type being used for the IDIR user search, but it wasn't a good design. Replaced it with more type-safe functions, which also might not be a good design.

* feat(ui): TMS-2509 rbac for management

Added RBAC control over editing tenants (tenant owner) and managing users (user admin).