feat: TMS-2103 request new tenant (#131)
* refactor(ui): TMS-2103 alphabetize css

Did a little cleanup of the global css so that things are easier to find.

* feat(ui): TMS-2103 use loading wrapper

Use the loading wrapper for the list of tenants.

* feat(ui): TMS-2103 change create button

Change the Create New Tenant button to be Request New Tenant.

* feat(ui): TMS-2103 tenant create dialog

Rework the tenant create dialog to be tenant request. It is trending to be very similar to the details form.

* fix(ui): TMS-2103 description of tenant

Change label to be "Description of Tenant" for consistency.

* feat(ui): TMS-2103 tenant request call

Added the tenant request API call, the store function to call it, and then used them in the view component.