fix(ui): TMS-2519 fixed bugs and renamed files (#133)
* fix(ui): TMS-2519 notification updates

Updated the titles and messages to more closely match the wireframes.

* fix(ui): TMS-2519 owner is user admin

Updated the logic so that users with the tenant owner role automatically get user admin - even if they don't have user admin.

* fix(ui): TMS-2519 dialog emits error

Vue was complaining about the generic emit that was matching the button text. Switched to a permanent emit event with a generic parameter.

* reafactor(ui): TMS-2519 rename tenant subtype

Renamed TenantEditFields to TenantDetailFields

* reafactor(ui): TMS-2519 rename tenant dialog

Renamed TenantCreateDialog to TenantrequestDialog