fix(ui): TMS-2519 miscellaneous bug fixes (#134)
* refactor(ui): TMS-2519 rename tenant subtype

Missed one renaming of  TenantEditFields to TenantDetailFields

* fix(ui): TMS-2519 everyone gets bcsans

Move the global BCSans config into the CSS as it wasn't being picked up from the vuetify.ts

* fix(ui): TMS-2519 two character user search

The search for IDIR users must have at least two characters, so enforce that in the UI. Could still be improved as there is no user feedback.

* feat(ui): TMS-2519 feedback when no results

Fixed up the user search so that the alert is displayed if there are no matching results. Using null as no search and [] as no results.

* feat(ui): TMS-2519 data table stripes

Added even-row background stripes to the data tables. This involved an update of the vuetify dependency.