fix(ui): TMS-2519 final set of ui bugs (#136)
* fix(ui): TMS-2519 handle tenant load

Update the tenant manager to clear the spinner when the tenant load fails.

* fix(ui): TMS-2519 disabled text colour

Remove the setting for the disabled text colour, as Vuetify is going to use opacity or some other way of lightening it. Just leave it dark and the lightened text will have enough contrast.

* fix(ui): TMS-2519 user sort change

There was a bug that the user-selected sorting of the user search table would change when a user was selected.

* fix(ui): TMS-2519 roles display

The bug was that if a searched user is selected then the roles display - if a new search is done the roles weren't disappearing or clearing.

* fix(ui): TMS-2519 tenant details errors

The UI was updating the store with new tenant details before making the API call. This was causing problems when the call failed.