feat(ui): TMS-2244 add user during group creation (#156)
* feat(ui): TMS-2244 add user to group

First commit to do the work when the "add user" checkbox is clicked.

* feat(ui): TMS-2244 add user to group

Finish off the ticket by adding the store code and doing better error responses. Also update the validation message that is shown if there is a group name collision.

* fix(ui): TMS-2244 small typo

Removed extra space added by cut and paste.