refactor(ui): TMS-2587 setups and user container (#169)
* refactor(ui): TMS-2587 add user container

Adding a user management container so that the tenant management container doesn't need to be so complicated.

* refactor(ui): TMS-2587 setup consistency

Updated the component setup code so that it's organized in a consistent way.

* refactor(ui): TMS-2587 change dialog location

Moved the tenant request dialog to be in the tenantrequest directory.

* refactor(ui): TMS-2587 setup consistency

Updated the component setup code so that it's organized in a consistent way.

* refactor(ui): TMS-2587 setup consistency

Updated the component setup code so that it's organized in a consistent way.