build(dep): TMS-2646 keycloak and dependabot (#177)
* build(dep): TMS-2646 update keycloak

The SSO team has updated their systems and recommend the use of keycloak 26.2.0.

* build(dep): TMS-2646 update start-server-and-test

Dependabot identified vulnerabilities with the dependencies of start-server-and-test. Updated it to latest.

* build(dep): TMS-2646 update cypress

Dependabot identified vulnerabilities with the dependencies of Cypress. Updated it to latest.

* build(dep): TMS-2646 update eslint

NPM identified low-severity vulnerabilities with the dependencies of ESLint. Updated it to latest.

* build(dep): TMS-2646 update brace-expansion

NPM identified low-severity vulnerabilities with the dependencies of ESLint. The update of ESLint to latest did not pick up the newer version, so updated it in the lock file.

* build(dep): TMS-2646 update form-data

Dependabot identified a vulnerability with the dependencies of Axios. Updated the dependency itself and left Axios as-is for now.