fix: TMS-2677 devcontainer tweaks (#186)
* fix: TMS-2677 devcontainer jre

The SonarQube extension is complaining that there is no JRE installed. Install the recommended v17 JRE and configure it in the settings.

* fix: TMS-2677 no install recommends

The GitHub tool Trivy recommends that --no-install-recommends is used when installing Dockerfile packages.

* fix: TMS-2677 dockerfile health check

The GitHub tool Trivy recommends that a healthcheck is added to all Dockerfiles, even the one for the devcontainer.

* feat: TMS-2677 db migrations on build

Uncomment the database migrations so that they run when the container is rebuilt.