build(deps): TMS-2693 update dependencies (#188)
* build(deps): TMS-2693 cypress update

Update Cypress to latest, as it has a security vulnerability in its dependency tmp.

* build(deps): TMS-2693 remove joi types

NPM was complaining that "This is a stub types definition. joi provides its own type definitions, so you do not need this installed."

* build(deps): TMS-2693 remove json2csv

NPM was complaining "Package no longer supported". It does not appear to be used for anything.